Separate routes by namespace instead of http method
