+++ app/factory.py
@@ -1,6 +1,7 @@
 from flask import Flask
-from app.routes.get import get
-from app.routes.post import post
+from app.routes.base import base
+from app.routes.auth import auth
+from app.routes.raffle import raffle
 
 import app.config as config
 
@@ -27,6 +28,7 @@ def create_app():
 
     register_error_handlers(app)
 
-    app.register_blueprint(get)
-    app.register_blueprint(post)
+    app.register_blueprint(base)
+    app.register_blueprint(auth, url_prefix='/auth')
+    app.register_blueprint(raffle, url_prefix='/raffles')
     return app
